Split more expressions, e.g., Mux tvals
Change SplitExpressions logic to split out more expressions, but not
trivial expressions. This will now split out tval and fvals from Mux
ops, but not create redundant ones for nodes and connects.

Add a test that this fixes the associated issue and that this doesn't
make SplitExpressions non-idempotent.

The Verilog emitter and clock legalization tests are modified to pass
with these changes.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>